refactor(js)!: run WASM calls on the caller's thread instead of a Worker
Remove the internal Worker (browser) / worker_threads (Node) indirection.
wasmClient now calls into the Emscripten module directly via ccall +
Asyncify, with no message-passing layer in between.

BREAKING CHANGE: img2num no longer offloads WASM execution to a background
thread automatically. Heavy operations (gaussianBlur, bilateralFilter,
kmeans, imageToSvg) now run on whichever thread calls them — this can
block the browser main thread/UI or the Node event loop during
processing. Consumers who need non-blocking behavior must now wrap
calls in their own Worker or worker_thread.
